+++ sdk-incubator-lib/src/androidTest/kotlin/cash/z/ecc/android/sdk/fixture/FiatCurrencyFixture.kt
@@ -3,5 +3,7 @@ package cash.z.ecc.android.sdk.fixture
 import cash.z.ecc.android.sdk.model.FiatCurrency
 
 object FiatCurrencyFixture {
-    fun new() = FiatCurrency.USD
+    const val USD = "USD"
+
+    fun new(code: String = USD) = FiatCurrency(code)
 }