+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Synchronizer.isValidTexAddr` which checks whether the given address is a valid ZIP 320 TEX address
 - `Synchronizer.exchangeRateUsd` to observe current USD exchange rate
 - `Synchronizer.refreshExchangeRateUsd` refreshes `Synchronizer.exchangeRateUsd` 
+- `ZatoshiExt.toFiatString` is now a public function
 
 ### Changed
 - Migrated to Rust 1.80.0